Move rayon specific code to it's own file
